[SSHD-1264] Use the same KEX proposal for every KEX in a session
The KEX proposal (algorithms and so on) is created once upon the first
KEX, and subsequently must not be changed anymore during the session.
Re-KEX must use the same proposal in this session, otherwise it may
fail because suddenly a different host key might be used (if the server
has several), or it might not be possible to negotiate algorithms when
it was possible on the initial KEX.

OpenSSH also evaluates the proposal only once per session and then
re-uses it; only the cookie in the SSH_MSG_KEX_INIT message changes
every time it is sent.
